+++ addOns/authhelper/CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
+### Added
+- If authentication fails then try to find a likely looking login link.
+
 ### Changed
 - Prefer form related fields in Browser Based Authentication for the selection of username field.
 - Tweaked the auth report summary keys.